Change Quick Start directions to use Google Earth Web
Google Earth Web now supports uploading KML without turning on feature in the settings. For advanced use it will still be better in the Google Earth Pro desktop application, but for first time users who do not have the application installed, it will be much easier to use the web app.